requirements.txt
google-generativeai==0.8.3
pandas==2.2.3
pypdf==5.1.0
google-cloud-translate==2.0.1
scikit-learn==1.6.0